refactor of backend util functions - Circular imports prevention
