+++ darwin/future/data_objects/team.py
@@ -1,7 +1,11 @@
-from typing import List, Optional
+from __future__ import annotations
+
+from typing import List, Optional, Tuple
 
 from pydantic import validator
 
+# from darwin.future.core.backend import get_team
+from darwin.future.core.client import Client
 from darwin.future.data_objects.dataset import DatasetList
 from darwin.future.data_objects.team_member_role import TeamMemberRole
 from darwin.future.data_objects.validators import parse_name
@@ -51,5 +55,32 @@ class Team(DefaultDarwin):
     # Data Validation
     _slug_validator = validator("slug", allow_reuse=True)(parse_name)
 
+    @staticmethod
+    def from_client(client: Client, team_slug: Optional[str] = None) -> Team:
+        """Returns the team with the given slug"""
+        if not team_slug:
+            team_slug = client.config.default_team
+        return get_team(client, team_slug)
+
 
 TeamList = List[Team]
+
+
+def get_team(client: Client, team_slug: Optional[str] = None) -> Team:
+    """Returns the team with the given slug"""
+    if not team_slug:
+        team_slug = client.config.default_team
+    response = client.get(f"/teams/{team_slug}/")
+    return Team.parse_obj(response)
+
+
+def get_team_members(client: Client) -> Tuple[List[TeamMember], List[Exception]]:
+    response = client.get(f"/memberships")
+    members = []
+    errors = []
+    for item in response:
+        try:
+            members.append(TeamMember.parse_obj(item))
+        except Exception as e:
+            errors.append(e)
+    return (members, errors)
