+++ darwin/future/tests/core/fixtures.py
@@ -3,7 +3,8 @@ from typing import List
 import pytest
 
 from darwin.future.core.client import Client, DarwinConfig
-from darwin.future.data_objects.team import Team, TeamMember, TeamMemberRole
+from darwin.future.data_objects.team import Team, TeamMember
+from darwin.future.data_objects.team_member_role import TeamMemberRole
 
 
 @pytest.fixture
