+++ darwin/future/tests/data_objects/test_team.py
@@ -4,9 +4,8 @@ import pytest
 import responses
 from pydantic import ValidationError
 
-from darwin.future.core import backend as be
 from darwin.future.core.client import Client
-from darwin.future.data_objects.team import Team, TeamMember
+from darwin.future.data_objects.team import Team, TeamMember, get_team, get_team_members
 from darwin.future.tests.core.fixtures import *
 from darwin.future.tests.fixtures import *
 
@@ -17,7 +16,7 @@ def test_get_team_returns_valid_team(base_client: Client, base_team_json: dict, 
     with responses.RequestsMock() as rsps:
         rsps.add(responses.GET, endpoint, json=base_team_json)
 
-        team = be.get_team(base_client, slug)
+        team = get_team(base_client, slug)
         assert team == base_team
 
 
@@ -28,7 +27,7 @@ def test_get_team_fails_on_incorrect_input(base_client: Client, base_team: Team)
         rsps.add(responses.GET, endpoint, json={})
 
         with pytest.raises(ValidationError):
-            team = be.get_team(base_client, slug)
+            team = get_team(base_client, slug)
 
 
 def test_get_team_members_returns_valid_list(base_client: Client, base_team_member_json: dict) -> None:
@@ -37,7 +36,7 @@ def test_get_team_members_returns_valid_list(base_client: Client, base_team_memb
     with responses.RequestsMock() as rsps:
         rsps.add(responses.GET, endpoint, json=[base_team_member_json, base_team_member_json])
 
-        members, errors = be.get_team_members(base_client)
+        members, errors = get_team_members(base_client)
         assert len(members) == 2
         assert len(errors) == 0
         assert members == synthetic_list
@@ -48,8 +47,20 @@ def test_get_team_members_fails_on_incorrect_input(base_client: Client, base_tea
     with responses.RequestsMock() as rsps:
         rsps.add(responses.GET, endpoint, json=[base_team_member_json, {}])
 
-        members, errors = be.get_team_members(base_client)
+        members, errors = get_team_members(base_client)
         assert len(members) == 1
         assert len(errors) == 1
         assert isinstance(errors[0], ValidationError)
         assert isinstance(members[0], TeamMember)
+
+
+def test_team_from_client(base_client: Client, base_team_json: dict, base_team: Team) -> None:
+    with responses.RequestsMock() as rsps:
+        rsps.add(
+            responses.GET,
+            base_client.config.api_endpoint + f"teams/{base_client.config.default_team}",
+            json=base_team_json,
+        )
+
+        team = Team.from_client(base_client)
+        assert team == base_team
